Calculate number of selected months in main app (#33)
Pass it to the module.
The error occured when e.g. the expenses for a given category didn't start until the third month. Then the inside-module reactive would count two months less than the selected span.
Now it's count at root level so it has access to the datespan selectors, and passes the number of months to the module.